jobs/bulkload/requirements.txt
psycopg2-binary
cuid2
duckdb